#10 : 3 more buildings ! - Implemented Canon Tower and its upgrade. It shoots ballistic projectiles that follow a 3-point Bezier curve to the ground and explodes on impact. - Implemented Boulder, a cheap obstacle with low health. - Added Projectile Death Effect script which will handle all effects that trigger on impact. Current implementation : explode. - Simplified Building directory structure - Heavily refactored TowerAttack and ProjectileMovement such that these two scripts are sufficient to handle any projectile type through their parameters and public variables. Simplified coding logic.

+++ COMP-476-Project/Assets/Scripts/Enemies/EnemyAttributes.cs
@@ -5,9 +5,9 @@ using UnityEngine.UI;
 
 public class EnemyAttributes : MonoBehaviour
 {
-    public int health;
-    int maxHealth;
-    public int damage;
+    public float health;
+    float maxHealth;
+    public float damage;
     public float speed;
     public bool isDead;
     public float heightOffset;
@@ -34,12 +34,12 @@ public class EnemyAttributes : MonoBehaviour
         return isDead;
     }
 
-    public int GetHealth()
+    public float GetHealth()
     {
         return health;
     }
 
-    public void DealDamage(int dmg)
+    public void DealDamage(float dmg)
     {
         health -= dmg;
         if (health <= 0)
